Extract getResultSeverity() into separate file
