+++ lib/index.js
@@ -8,32 +8,9 @@ const path = require('path');
 const escapeStringRegexp = require('escape-string-regexp');
 
 const isString = require('./is-string');
+const getResultSeverity = require('./get-result-severity');
 
 const BUILD_DIR_REGEXP = new RegExp(`(${escapeStringRegexp(path.sep)})?lib(${escapeStringRegexp(path.sep)})?$`);
-
-/**
- * Calculates the severity of a eslint.linter.verify result
- * @param {Array} resultMessages ESLint's verify() result array
- *    @see: http://eslint.org/docs/developer-guide/nodejs-api#linter
- *
- * @returns {Number} accumulatedSeverity The total severity from of the list of results
- *    0 indicates all-clear
- *    1 indicates a warning-level result
- *    > 1 indicates an error-level result
- */
-function getResultSeverity() {
-  let resultMessages = arguments[0] || [];
-
-  return resultMessages.reduce((accumulatedSeverity, message) => {
-    const severity = message.severity || 0;
-
-    if (message.fatal || severity === 2) {
-      return accumulatedSeverity + 2;
-    }
-
-    return accumulatedSeverity + severity;
-  }, 0);
-}
 
 function resolveInputDirectory(inputNode) {
   if (typeof inputNode === 'string') {
